Added support for an lsxcommand alias file to be auto-generated, and a "don't forcibly kill litestep" option in settings.ini
git-svn-id: http://code.google.com/archive/p/litestepportable/source/trunk@173 38ca8cbe-b76d-4e91-b741-e71bb1f618b7

+++ Other/Source/ShellSwitcher/Switcher.cpp
@@ -63,8 +63,12 @@ int PromptAction() {
 
 bool KillLitestep() {
      // Terminating Litestep
-     HANDLE LitestepHandle = OpenProcess(PROCESS_ALL_ACCESS, FALSE, globalLitestepPID);
-     TerminateProcess(LitestepHandle, EXIT_SUCCESS);
+     if (GetBoolValue("Force Litestep Termination")) {
+         HANDLE LitestepHandle = OpenProcess(PROCESS_ALL_ACCESS, FALSE, globalLitestepPID);
+         TerminateProcess(LitestepHandle, EXIT_SUCCESS);
+     } else {
+         ShellExecute( NULL, "open", GetValue("Litestep").c_str(), "!quit", NULL, SW_SHOW);
+     }
     return false;
 }
